Fix several classes of compiler diagnostic & make fatal
 * -Wno-nested-externs
 * -Wno-strict-prototypes
 * -Wno-double-promotion
 * -Wno-sign-compare

+++ ports/raspberrypi/common-hal/rp2pio/__init__.c
@@ -33,7 +33,7 @@ bool common_hal_rp2pio_pins_are_sequential(size_t len, mp_obj_t *items) {
         return true;
     }
     const mcu_pin_obj_t *last_pin = validate_obj_is_pin(items[0]);
-    for (int i = 1; i < len; i++) {
+    for (size_t i = 1; i < len; i++) {
         const mcu_pin_obj_t *pin = validate_obj_is_pin(items[i]);
         if (pin->number != last_pin->number + 1) {
             return false;
